Move funcs from scooped_view to record.go
This is the first step in a major refactoring of
the codebase. Scooped_view's functions all acted
on data in record. In order to do this, a useless
interface was made to make it easier for the
helper functions. The plan is to remove this
interface and have the models take more responsibility
for handling data.