Refactor ResolveReference and more cases in EntriesForDataCriteria
ResolveReference was iterating over all the entries in record
when it could immediately resolve which group of entries it
needed to search through. This refactoring then removed the need
for EntriesMap.
Also in this commit is some shortening of the
logic in the second switch statement in EntriesForDataCriteria.
This allows the code to be reasoned about more easily.