Redesign access to global maps
Instead of needing to insure the global maps were initialized before
touching them across the codebase, the hqmf and qrda maps now get
initialized once in the init function for the package in models.go.
Then outside packages can gain access to the struct that holds these
maps through the NewHds function.
For the value set map, it seems like the data can be more fluid. In
order to handle this, a NewValueSetMap function was created in
value_set.go that takes value sets that need to go into this map.
Then the function returns a value set map with the given value sets
in it.