Refactor cat1 view helper (#66)
* Move EntryInfosForPatient to record

This was a very large move because of the certain global maps
made in the exporter package. Most of cat1_view_helper.go and
hqmf_template_resolver.go had to be moved to the models package
and was dumped into the record.go file for simplicity. The next
stage of this refactor will be moving this logic out of models
and into its own package that will then act as a service for any
library to use for Health Data Standard specific look ups,
transformations, etc.

* Redesign access to global maps

Instead of needing to insure the global maps were initialized before
touching them across the codebase, the hqmf and qrda maps now get
initialized once in the init function for the package in models.go.
Then outside packages can gain access to the struct that holds these
maps through the NewHds function.
For the value set map, it seems like the data can be more fluid. In
order to handle this, a NewValueSetMap function was created in
value_set.go that takes value sets that need to go into this map.
Then the function returns a value set map with the given value sets
in it.