Remove codeToDisplay()
This function is not necessary since all it does is get the Entry
from the struct that satisfies the interface and call the Entry's
GetCodeDisplay() method. This can be called directly on the struct
that satisfies the interface because they all have Entry embedded
in them, thus they are Entries and have that method available.