Remove pebbe/util as a dependency in exporter
All pebbe/util.CheckErr did was log.Fatal so I replaced it with the
appropriate check using the standard library's log package.

Add log settings to package

The init function in exporter.go configures some settings for the log package
so that we get the filename and line number added with any log messages.