Update and clean escape function
Escape was responsible for too much and the hint was that it had to
take the empty interface as a parameter. Now it only acts as a helper
for templates to xml escape strings. Now there is a new function,
i64ptrToString that handles the responsibility of being a helper for
templates that need to print unix times, represented as *int64 in the code.

Escape was also using xml.Escape to escape strings. The Go team keeps
xml.Escape for compatibility reasons, but xml.EscapeText should be used
instead, so this has been updated.

As for removing the cases where escape was passing integers through an
xml escape function, integers cannot contain anything to be xml escaped,
so this was removed. The reasoning for this behavior was that the escape
function was serving a secondary purpose of dereferencing *int64s and
converting them to strings. This functionality has been moved to
i64ptrToString in order to be cleaner.