Add initial refactor to MVVM for exporter.
There is now a subpackage inside of exporter called document. This initial
refactor started with the allergy entry where a view model for allergy was
created with the original allergy model embedded in it. Then, the cat1 template
for allergy was added to the view model so that the view model could represent
itself in a cat1 form.