Move data from json files into byte arrays
I hope for this to be the first step towards not reading files at run time to
retrieve static data. This move is literally only one step from where we were
by moving the the data into byte array variables in a fixtures go package. This
allows us to now just unmarshal the data into our structures instead of having
to read the file beforehand which avoids potential errors with missing files,
etc.