Change Makefile to run go-bin-data with the -no-compress option
After doing an initial benchmark on the GenerateCat1 function, 50% of the
garbage created is because go-bindata gzips the templates before putting
them into templates.go and therefore has to unzip them before they can be
used. This happens on every call to any template. Simply giving go-bindata
the `-no-compress` option stops this and instantly decreases our memory
usage by about 62%. This is also helpful on computation, because 50% of cpu
time was spent collecting garbage.