Add ExporterCat1Init to library API
This init function is being added to hopefully avoid having to
unmarshal the measure and value set data every time a patient gets
exported. The usual use case is that many patients need to be
exported under the same measure with the same value sets.

This function unmarshals the measure and value set data into
structures the Go library uses to properly export the patients. This
way, the Go library can reuse these structures for all the patients
and take pressure off the GC.

This means that this function is expected to be called before any
calls to GenerateCat1 and to only be called once per need to change
the measure and value set data.

This new function also uses this opportunity to parse the templates it
will need to generate a cat1 file. This is another move for speed reasons
and could even be refactored to occur when the library is loaded, if some
changes were made to the way Cypress interacts with this library.