Add LoadMeasuresAndValueSets to library API
This function is being added to avoid having to unmarshal the
measure and value set data every time a patient gets exported.
The usual use case is that many patients need to be exported
under the same measures with the same value sets.

This function unmarshals the measure and value set data into
structures the Go library uses to properly export the patients. This
way, the Go library can reuse these structures for all the patients
and take pressure off the GC.

This means that this function is expected to be called before any
calls to GenerateCat1 and to only be called once per need to change
the measure and value set data.