Change ChechErr calls to panics with err.Error()
This also improves catchPanics to include the panic message.

CheckErr will call log.Fatal if the err is not nil. This will
shut down the worker that runs cdatools and there is no way to
recover from it. At least with panics, the state can be restored and
we can introduce better error handling systematically.